requirements.txt
django
djangorestframework
djangorestframework-jwt
markdown
django-filter
djangorestframework-simplejwt

